Storage modifications & DMP Export
Data matrices are now saved as JSON files in the Minio file system. The
storage of the JSON strings in the MYSQL database has been removed,
because the data was too large and storing such large files in a
database is not performing well.

DataTables on the import and matrix pages now work completely
asynchronously. This increases the performance of large tables
enormously.

Preparation for DMP export. Test buttons have been added to dmp.jsp.
Export functions for each funder will be added soon.